Convert duplicate items into single items that can be sent multiple times
For each item where there are multiple copies, there is now only a
single Archipelago item, but multiple copies are added to the item pool.

Locally placed item IDs are now placed starting from the maximum
Psychonauts ID for that item and decrease by one for each item placed.
Conversely, non-locally received items start from the minimum
Psychonauts ID for that item and the ID is increased by one for each
non-local item received.

To ensure that non-locally received items cannot use the Psychonauts ID
used by locally placed items, the Psychonauts IDs for all locally placed
items are calculated after the client connects to the server. This is
done by sending a LocationScouts request of all the local locations and
then running the same code that is used during PsychoRando seed
generation to calculate the Psychonauts IDs for all the locally placed
items.

Through normal gameplay, it is not possible to run out of Psychonauts
IDs for a specific item, because there are only as many placed into the
multiworld as exist, however, server cheat commands can send items
directly. In this case, it is possible to run out of Psychonauts IDs, in
which case, the client will print a warning that the item could not be
received when Psychonauts has run out of IDs.

When connecting to an Archipelago server, all previously received items
are re-sent to the client. This usually happens before the LocationInfo
response to the client sending a LocationScouts request for all local
locations. This means that initially received items must be temporarily
stored until the LocationInfo response is received.

PsychoSeed.gen_psy_ids() has been made more generic so that it can also
be called using LocationInfo scouted by the client. Additionally,
gen_psy_ids() now pre-sorts the input by location to ensure that the
same result is produced even if the locations are provided in a
different order because it is not clear if there are guarantees to the
order of the LocationInfo in response to a LocationScouts request.

Due to the increased complexity of receiving items, new functions have
been created for receiving local or non-local items specifically.

Most of the item groups have been removed because they are now
unnecessary.

Due to the use of `import *`, __init__.py was importing `Item` from
 `BaseClasses` indirectly from `.Items`. This import has been moved to
 __init__.py where it is actually used.