Add support for plando adding more items to local locations than can exist
Attempting to place an item locally when all the Psychonauts IDs for
that item have been used up will instead place an AP placeholder item.

Because of the fully remote item handling, collecting that AP
placeholder item will cause the server to tell the client that it has
received the item. The client can then send Psychonauts the correct item
that couldn't be placed locally due to the ID restrictions.

This also means that PsiCards can now be safely used as filler items
without the possibility of running out of IDs. Though, because the
number of locations currently matches the size of the item pool, these
shouldn't be used.